test(progress-view): fold ChatExport suites into one per-module suite
Delete ChatExportController.vitest.ts, a near-verbatim mocked copy of
loadChatExportInput.vitest.ts re-testing the shared loader through the
thin buildExportInput wrapper, and drop the three exportAsHtml failure
tests (plus expectStreamLogsMissing) whose assembleTrace decisions are
owned by traceAssembler.vitest.ts. Rename ChatExportControllerHtml to
ChatExportController (one suite per module) and pin conversation_missing
through the wrapper with one real-storage test; config_missing
propagation stays pinned by the suite's first test.
