+++ src/test-kernel/progressView/ChatExportController.vitest.ts
@@ -1,121 +1,171 @@
-/* eslint-disable import/order -- Vitest mocks must be declared before importing the module under test. */
-import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { writeFile } from 'node:fs/promises';
+import * as path from 'node:path';
 
+import { beforeEach, describe, expect, it } from 'vitest';
+
+import { getExecutionStore } from '@agent/storage';
 import type { AgentConfig } from '@agent/core/definition/AgentConfig';
-import { DEFAULT_TOOL_CONFIG } from '@shared/schemas';
-
-const mocks = vi.hoisted(() => ({
-  readConfig: vi.fn(),
-  readConversation: vi.fn(),
-  readMeta: vi.fn(),
-}));
-
-vi.mock('@agent/storage', () => ({
-  getExecutionStore: vi.fn(() => ({
-    readConfig: mocks.readConfig,
-    readMeta: mocks.readMeta,
-  })),
-}));
-
-vi.mock('@transcript', () => ({
-  hasCompletedRunConversationEvidence: vi.fn(
-    ({ conversation }: { conversation: unknown[] | null }) =>
-      (conversation?.length ?? 0) > 0,
-  ),
-  readCompletedRunConversation: vi.fn(async () => {
-    const conversation = await mocks.readConversation();
-    return {
-      conversation,
-      source: conversation === null ? 'none' : 'streamLog',
-    };
-  }),
-}));
-
-// Imported after vi.mock so the mocked dependency is in place.
+import { getStreamTabId } from '@agent/runtime/streamTab';
 import { ChatExportController } from '@controllers/progressView/ChatExportController';
+import { MemoryStateStore } from '@platform/defaults/memoryState';
+import { nodeFilesystem } from '@platform/defaults/nodeFilesystem';
+import { createNodeWorkspace } from '@platform/defaults/nodeWorkspace';
+import { WorkspaceStorageProvider } from '@platform/defaults/workspaceStorage';
+import {
+  LOG_LEVELS,
+  MESSAGE_TYPES,
+  STREAM_LOG_ENTRY_TYPES,
+  AgentCategory,
+  DEFAULT_TOOL_CONFIG,
+} from '@shared/schemas';
+import type { ExecutionId, StreamTabId } from '@shared/schemas';
+import { createFakePlatform } from '@test/support/FakePlatform';
+import { makeTempDir, useTempDirs } from '@test/support/tempDirPlatform';
+import { appendTranscriptEntry } from '@test/support/storeTestDrivers';
+import { StreamLogStore } from '@transcript';
+import { StorageFS } from '@utils/files/storageFS';
 
-const config = {
-  agent: 'correct',
-  model: 'deepseekT',
-  instruction: 'Polish the introduction.',
-  agentCategory: 'workflow',
-  inputFiles: ['chapters/intro.tex'],
-  outputFiles: ['chapters/intro.tex'],
-  contextFiles: [],
-  mediaFiles: [],
-  editedFile: null,
-  editedFiles: [],
-  memories: [],
-  toolConfig: DEFAULT_TOOL_CONFIG,
-} as AgentConfig;
+const TEMPLATE =
+  '<!doctype html><html><head><title>t</title>' +
+  '<script type="module" crossorigin src="./index.js"></script>' +
+  '</head><body></body></html>';
+
+const tempDirs = useTempDirs();
+
+async function installStoragePlatform(): Promise<void> {
+  const tempDir = await makeTempDir('texra-html-export-', tempDirs);
+  const workspaceDir = path.join(tempDir, 'workspace');
+  const storageRoot = path.join(tempDir, 'storage');
+  const { initPlatform } = await import('@platform/platform');
+  initPlatform(
+    createFakePlatform(
+      { workspacePath: workspaceDir },
+      {
+        fs: nodeFilesystem,
+        workspace: createNodeWorkspace(() => workspaceDir),
+        storage: new WorkspaceStorageProvider(storageRoot, workspaceDir),
+        globalState: new MemoryStateStore(),
+        workspaceState: new MemoryStateStore(),
+      },
+    ),
+  );
+}
+
+async function writeTemplate(): Promise<string> {
+  const tempDir = await makeTempDir('texra-template-', tempDirs);
+  const templatePath = path.join(tempDir, 'index.html');
+  await writeFile(templatePath, TEMPLATE, 'utf8');
+  return templatePath;
+}
+
+function config(overrides: Partial<AgentConfig> = {}): AgentConfig {
+  return {
+    inputFiles: [],
+    contextFiles: [],
+    mediaFiles: [],
+    outputFiles: [],
+    editedFile: null,
+    agent: 'orchestrator',
+    model: 'deepseekT',
+    instruction: 'Solve the problem.',
+    agentCategory: AgentCategory.ToolUse,
+    editedFiles: [],
+    toolConfig: DEFAULT_TOOL_CONFIG,
+    memories: [],
+    workingDirectory: '/workspace',
+    cliOutputFile: null,
+    cliMultiAgentPresetId: null,
+    ...overrides,
+  };
+}
+
+async function persistTranscriptEntry(
+  executionId: ExecutionId,
+  agent: string,
+): Promise<StreamTabId> {
+  const streamId = getStreamTabId(agent, { executionId });
+  const store = await StreamLogStore.open();
+  appendTranscriptEntry(store, streamId, {
+    id: 'entry-1',
+    type: STREAM_LOG_ENTRY_TYPES.LOG,
+    level: LOG_LEVELS.INFO,
+    timestamp: 100,
+    messageType: MESSAGE_TYPES.DEFAULT,
+    text: 'hello',
+  });
+  await store.flush();
+  return streamId;
+}
+
+describe('ChatExportController.exportAsHtml', () => {
+  const controller = new ChatExportController({ latexPreamble: '' });
+
+  beforeEach(installStoragePlatform);
+
+  it('returns config_missing when nothing is stored', async () => {
+    const templatePath = await writeTemplate();
+
+    const outcome = await controller.exportAsHtml('missing', templatePath);
+
+    expect(outcome).toEqual({ status: 'config_missing' });
+  });
+
+  it('writes a self-contained HTML file with the trace embedded, when everything is present', async () => {
+    const templatePath = await writeTemplate();
+    const executionId = 'exec-full' as ExecutionId;
+    const executionConfig = config({ agent: 'review', model: 'sonnet46T' });
+    await getExecutionStore(executionId).writeRunRecord(executionConfig);
+    const streamId = await persistTranscriptEntry(executionId, 'review');
+    await getExecutionStore(executionId).writeMeta({
+      timestamp: '2026-07-05T00:00:00.000Z',
+      outcome: 'completed',
+      streamId,
+    });
+
+    const outcome = await controller.exportAsHtml(executionId, templatePath);
+
+    expect(outcome.status).toBe('ok');
+    if (outcome.status !== 'ok') return;
+    expect(outcome.result.storagePath).toMatch(
+      /^executions\/exec-full\/texra-chat-.*\.html$/,
+    );
+
+    const written = await nodeFilesystem.readFile(
+      StorageFS.fullPath(outcome.result.storagePath),
+    );
+    const html = new TextDecoder().decode(written);
+    expect(html).toContain('<script>window.__TEXRA_TRACE__');
+    expect(html).toContain('"id":"entry-1"');
+    expect(html).toContain(
+      '<script type="module" crossorigin src="./index.js">',
+    );
+  });
+
+  it('throws when the standalone template bundle is missing', async () => {
+    const executionId = 'exec-missing-template' as ExecutionId;
+    await getExecutionStore(executionId).writeRunRecord(config());
+    const streamId = await persistTranscriptEntry(executionId, 'orchestrator');
+    await getExecutionStore(executionId).writeMeta({
+      timestamp: '2026-07-05T00:00:00.000Z',
+      streamId,
+    });
+
+    await expect(
+      controller.exportAsHtml(executionId, '/nonexistent/index.html'),
+    ).rejects.toThrow(/Trace-viewer standalone bundle missing/);
+  });
+});
 
 describe('ChatExportController.buildExportInput', () => {
   const controller = new ChatExportController({ latexPreamble: '' });
 
-  beforeEach(() => {
-    vi.clearAllMocks();
-    mocks.readConfig.mockResolvedValue(null);
-    mocks.readConversation.mockResolvedValue(null);
-    mocks.readMeta.mockResolvedValue(null);
-  });
+  beforeEach(installStoragePlatform);
 
-  it('builds export input from the stored config, conversation, and meta', async () => {
-    mocks.readConfig.mockResolvedValue(config);
-    mocks.readConversation.mockResolvedValue([
-      { role: 'user', content: 'Polish the lemma.' },
-      { role: 'assistant', content: 'Done.' },
-    ]);
-    mocks.readMeta.mockResolvedValue({
-      timestamp: '2026-05-18T08:00:00.000Z',
-      description: 'Polish pass',
-    });
+  it('reports conversation_missing when a config is stored but no transcript exists', async () => {
+    const executionId = 'exec-no-chat' as ExecutionId;
+    await getExecutionStore(executionId).writeRunRecord(config());
 
-    const result = await controller.buildExportInput('a1');
-
-    expect(result).toEqual({
-      status: 'ok',
-      exportInput: {
-        timestamp: '2026-05-18T08:00:00.000Z',
-        description: 'Polish pass',
-        config: {
-          agent: 'correct',
-          model: 'deepseekT',
-          instruction: 'Polish the introduction.',
-          inputFiles: ['chapters/intro.tex'],
-          mediaFiles: [],
-          contextFiles: [],
-          outputFiles: ['chapters/intro.tex'],
-        },
-        messages: [
-          { role: 'user', content: 'Polish the lemma.' },
-          { role: 'assistant', content: 'Done.' },
-        ],
-      },
-    });
-  });
-
-  it('reports "config_missing" when nothing is stored at all', async () => {
-    await expect(controller.buildExportInput('missing')).resolves.toEqual({
-      status: 'config_missing',
-    });
-  });
-
-  it('reports "conversation_missing" when config exists but conversation does not', async () => {
-    mocks.readConfig.mockResolvedValue(config);
-
-    await expect(controller.buildExportInput('a1')).resolves.toEqual({
-      status: 'conversation_missing',
-    });
-  });
-
-  it('reports "conversation_missing" (not "ok") when the stored conversation is only an empty array', async () => {
-    // A stored-but-empty conversation array is truthy — this must not be
-    // mistaken for "a conversation is present" (matches the CLI's
-    // not_found/incomplete reconciliation for the same underlying case).
-    mocks.readConfig.mockResolvedValue(config);
-    mocks.readConversation.mockResolvedValue([]);
-
-    await expect(controller.buildExportInput('a1')).resolves.toEqual({
+    await expect(controller.buildExportInput(executionId)).resolves.toEqual({
       status: 'conversation_missing',
     });
   });
